better db:reset bugfix

+++ Rakefile
@@ -21,10 +21,22 @@ namespace :db do
   task :nuke do |cmd, args|
     puts "Nuking the database."
 
+
+    ## this works but it's hacky ##
     DB.tables.each do |table|
-      DB.run("DROP TABLE #{table}")
+      binding.pry
+      if table != :playlists
+        DB.run("DROP TABLE #{table}")
+      end
     end
+    DB.run("DROP TABLE #{:playlists}")
   end
+
+  ## working on better solution:
+
+  # DB.schema(:voters).detect do |column| column.include?(:song_id) end
+  
+  ##
  
   desc "Reset the database"
   task :reset => [:nuke, :migrate]
